Factor out database checks into a separate module.
